Adds user check for workflow execution
Prevents workflow execution for external users.
It checks if the user is a member of the organization
before proceeding with the build. Also enables execution
if a pull request has a "safe to test" label or the ref is a tag.
This adds an extra layer of security and control over the
CI/CD pipeline.