Adds watchdog timer support to CIP objects
Adds support for a watchdog timer device to the CIP technology object.

This change introduces a new recipe parameter (RV_WATCHDOG) that, when set, associates a watchdog device with the CIP object. The CIP object then checks the watchdog status in its error checking routine (_CheckErr). If the watchdog is not "kicked", an ERR_WATCHDOG error is returned. A new test case is added to verify the functionality.