Relocates lua shared library search path
Updates the CMakeLists.txt file to dynamically determine the location of the lua shared library.

This change ensures that the demo projects can correctly locate and load the shared library, even when the build directory structure changes.

It also replaces PROJECT_BINARY_DIR with TARGET_FILE to correctly resolve the path where the lua executable can be found.