[test] Adds a destructor to reset the time in `DeltaMilliSecSubHooker` (#1122)
* [test] Adds a destructor to reset the time

Adds a destructor to the `DeltaMilliSecSubHooker` class to ensure that the default time is restored when the object goes out of scope.

* Adds deleted assignment operator to class

Adds the deleted assignment operator to the class to prevent accidental copying.
This ensures that the class behaves as intended and avoids potential issues related to unintended object duplication.

---------

Co-authored-by: AlexandrBehunkov <asutp10@pda.savushkin.by>