Controls extra step message output
Modifies the `on_extra_step` function to provide more control
over when the step time is included in the debug output. This
enhances the clarity and relevance of the debug messages.

Removes the default value for `step_time` in the primary `on_extra_step`
function to allow the `is_print_time` to control the output.

Adds a overload for the `on_extra_step` function that takes only the `step_idx`
as argument that calls the main `on_extra_step` function with a default value
for `step_time` of 0 and `is_print_time` to false.